# openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HostedEvaluationRequest": {
        "additionalProperties": false,
        "properties": {
          "acoustic_confidence_scores": {
            "anyOf": [
              {
                "items": {
                  "maximum": 1.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                "maxItems": 256,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acoustic Confidence Scores"
          },
          "action": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Action",
            "type": "string"
          },
          "attempt": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Attempt",
            "type": "string"
          },
          "comparison_mode": {
            "default": false,
            "title": "Comparison Mode",
            "type": "boolean"
          },
          "evaluation_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Id"
          },
          "target": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target",
          "attempt",
          "action"
        ],
        "title": "HostedEvaluationRequest",
        "type": "object"
      },
      "PatternCategory": {
        "description": "Interpretation of a child's attempt. These are not clinical diagnoses.",
        "enum": [
          "clean_reading",
          "developmental_speech_pattern",
          "decoding_error",
          "uncertain_requires_review"
        ],
        "title": "PatternCategory",
        "type": "string"
      },
      "ProfileLookupRequest": {
        "additionalProperties": false,
        "properties": {
          "user_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "ProfileLookupRequest",
        "type": "object"
      },
      "ProfileUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/PatternCategory",
            "default": "developmental_speech_pattern"
          },
          "is_correct": {
            "default": false,
            "title": "Is Correct",
            "type": "boolean"
          },
          "pattern": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Pattern",
            "type": "string"
          },
          "response_time": {
            "anyOf": [
              {
                "maximum": 3600.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Time"
          },
          "user_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "pattern"
        ],
        "title": "ProfileUpdateRequest",
        "type": "object"
      },
...
      "ReportRequest": {
        "additionalProperties": false,
        "properties": {
          "evaluations": {
            "items": {
              "$ref": "#/components/schemas/ReportEvaluationRequest"
            },
            "maxItems": 10000,
            "minItems": 1,
            "title": "Evaluations",
            "type": "array"
          }
        },
        "required": [
          "evaluations"
        ],
        "title": "ReportRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "paga-eval",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/evaluations": {
      "post": {
        "operationId": "evaluate_v1_evaluations_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostedEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Evaluate V1 Evaluations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate"
      }
    },
    "/v1/maintenance/prune": {
      "post": {
        "operationId": "prune_profiles_v1_maintenance_prune_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Prune Profiles V1 Maintenance Prune Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prune Profiles"
      }
    },
    "/v1/operations/stats": {
      "get": {
        "operationId": "operation_stats_v1_operations_stats_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Operation Stats V1 Operations Stats Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operation Stats"
      }
    },
    "/v1/profiles": {
      "post": {
        "operationId": "update_profile_v1_profiles_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Profile V1 Profiles Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile"
      }
    },
    "/v1/profiles/delete": {
      "post": {
        "operationId": "delete_profile_v1_profiles_delete_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileLookupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Profile V1 Profiles Delete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Profile"
      }
    },
    "/v1/profiles/lookup": {
      "post": {
        "operationId": "get_profile_v1_profiles_lookup_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileLookupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Profile V1 Profiles Lookup Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile"
      }
    },
    "/v1/reports": {
      "post": {
        "operationId": "report_v1_reports_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Report V1 Reports Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report"
      }
    }
  }
}